Create ImportMeshElement.
